requirements.txt
groq>=0.9.0
anthropic>=0.28.0
requests>=2.31.0
python-dotenv>=1.0.0
tweepy>=4.14.0
python-binance>=1.0.16
web3>=6.15.0
